+++ .vscode/extensions.json
@@ -3,6 +3,7 @@
         "editorconfig.editorconfig",
         "golang.go",
         "ms-azuretools.vscode-docker",
+        "ms-dotnettools.csharp",
         "ms-vscode.cpptools",
         "ms-vscode.makefile-tools",
         "rebornix.ruby",
