--- MandelbrotSet/csharp/.vscode/extensions.json
@@ -1,13 +1,0 @@
-{
-    // See http://go.microsoft.com/fwlink/?LinkId=827846 to learn about workspace recommendations.
-    // Extension identifier format: ${publisher}.${name}. Example: vscode.csharp
-    // List of extensions which should be recommended for users of this workspace.
-    "recommendations": [
-        "jchannon.csharpextensions",
-        "k--kato.docomment",
-        "ms-vscode.csharp",
-        "visualstudioexptteam.vscodeintellicode"
-    ],
-    // List of extensions recommended by VS Code that should not be recommended for users of this workspace.
-    "unwantedRecommendations": []
-}
