--- MandelbrotSet/csharp/MandelbrotSetCreator.sln
@@ -1,34 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{54D74E06-646C-452C-B1BD-F4B8F2356726}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataSource", "DataSource\DataSource.csproj", "{EA547746-0845-4FCC-8FCD-D2ABC89ECE55}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Executor", "Executor\Executor.csproj", "{CB92C837-32BE-42EC-A863-EB4F9F19BBBA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{54D74E06-646C-452C-B1BD-F4B8F2356726}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{54D74E06-646C-452C-B1BD-F4B8F2356726}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{54D74E06-646C-452C-B1BD-F4B8F2356726}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{54D74E06-646C-452C-B1BD-F4B8F2356726}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA547746-0845-4FCC-8FCD-D2ABC89ECE55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA547746-0845-4FCC-8FCD-D2ABC89ECE55}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA547746-0845-4FCC-8FCD-D2ABC89ECE55}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA547746-0845-4FCC-8FCD-D2ABC89ECE55}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CB92C837-32BE-42EC-A863-EB4F9F19BBBA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CB92C837-32BE-42EC-A863-EB4F9F19BBBA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CB92C837-32BE-42EC-A863-EB4F9F19BBBA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CB92C837-32BE-42EC-A863-EB4F9F19BBBA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
