+++ csharp/MandelbrotData/GraphParallelModel.cs
@@ -1,10 +1,8 @@
-using Core.Entities;
-using Core.Models;
-using System.Linq;
+using MandelbrotCore.Entities;
+using MandelbrotCore.Models;
 using System.Numerics;
-using System.Threading.Tasks;
 
-namespace DataSource
+namespace MandelbrotData
 {
     /// <summary>
     /// グラフ生成モデル(並列処理版)
