+++ csharp/MandelbrotExecutor/MandelbrotExecutor.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MandelbrotCore\MandelbrotCore.csproj" />
+    <ProjectReference Include="..\MandelbrotData\MandelbrotData.csproj" />
+  </ItemGroup>
+</Project>
